requirements.txt
python-telegram-bot==20.0
asyncio
python-dotenv==1.0.0
uvicorn==0.23.0
fastapi==0.98.0
starlette
